requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.39
psycopg[binary]==3.2.6
pydantic==2.10.6
pydantic-settings==2.8.1
python-jose[cryptography]==3.3.0
argon2-cffi==23.1.0
python-multipart==0.0.20
redis==5.2.1
alembic==1.15.1
